+++ pink/barriers/__init__.py
@@ -7,9 +7,14 @@
 """Control Barrier Functions."""
 
 from .barrier import Barrier
-from .exceptions import NoPositionLimitProvided, NegativeMinimumDistance
+from .exceptions import (
+    NoPositionLimitProvided,
+    NegativeMinimumDistance,
+    InvalidCollisionPairs,
+)
 from .position_barrier import PositionBarrier
 from .body_spherical_barrier import BodySphericalBarrier
+from .self_collision_barrier import SelfCollisionBarrier
 
 __all__ = [
     "Barrier",
@@ -17,4 +22,6 @@ __all__ = [
     "NoPositionLimitProvided",
     "NegativeMinimumDistance",
     "BodySphericalBarrier",
+    "SelfCollisionBarrier",
+    "InvalidCollisionPairs",
 ]