+++ pink/utils.py
@@ -112,3 +112,30 @@ class VectorSpace:
     def zeros(self) -> np.ndarray:
         """Zero vector of the space."""
         return self.__zeros
+
+
+def process_collision_pairs(
+    model: pin.Model, collision_model: pin.GeometryModel, srdf_path: str = ""
+) -> pin.GeometryData:
+    """Process collision pairs.
+
+    Args:
+        model: robot model.
+        collision_model: Collision model of the robot.
+        srdf_path: Path to the SRDF file, which used to exclude collision pairs.
+            Defaults to empty string, meaning no collision pairs are excluded.
+
+    Returns:
+        collision_data: Collision data, generated after updating collision_model.
+    """
+    collision_model.addAllCollisionPairs()
+    if srdf_path != "":
+        pin.removeCollisionPairs(model, collision_model, srdf_path)
+
+    # Collision models have been modified => re-generate corresponding data.
+    collision_data = pin.GeometryData(collision_model)
+
+    # Enable contact detection for avoiding Nans at collisions
+    collision_data.enable_contact = True
+
+    return collision_data